third_party/blink/web_tests/storage/indexeddb/object-lookups-in-versionchange-expected.txt

Regression test for http://webkit.org/b/102547

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "object-lookups-in-versionchange.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
db = event.target.result
transaction = event.target.transaction
store = db.createObjectStore('store')


Expecting exception from transaction.objectStore('no-such-store')
PASS Exception was thrown.
PASS code is DOMException.NOT_FOUND_ERR
PASS ename is 'NotFoundError'
Exception message: Failed to execute 'objectStore' on 'IDBTransaction': The specified object store was not found.


Expecting exception from store.index('no-such-index')
PASS Exception was thrown.
PASS code is DOMException.NOT_FOUND_ERR
PASS ename is 'NotFoundError'
Exception message: Failed to execute 'index' on 'IDBObjectStore': The specified index was not found.
PASS successfullyParsed is true

TEST COMPLETE
